fix(test): keep the key verifier importable on Node 18
CI Node 18 failed to LOAD tests/install-clients.test.ts: importing
wizard.ts pulls in @clack/prompts, and @clack/core imports `styleText`
from node:util, which does not exist until Node 20. The whole file died
before a single test ran — Node 20, 22 and Windows were all green, and so
was my local Node 24.

verifyApiKey moves to src/install/verify-key.ts, which has no interactive
dependencies, and the wizard imports it from there. Behaviour is
unchanged; the deadline is now reachable from a test on every supported
runtime.

Verified on Node 18.20.8 as well as 24.2: 673 tests, 665 pass, 0 fail.
